# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  domain.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  domain.RegisterRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  domain.ResetPasswordRequest:
    properties:
      password:
        minLength: 6
        type: string
      token:
        type: string
    required:
    - password
    - token
    type: object
...
  ecommerce-backend_internal_delivery_http.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/http.Error'
      message:
        type: string
      meta:
        $ref: '#/definitions/http.Meta'
      success:
        type: boolean
    type: object
  http.Error:
    properties:
      code:
        type: string
      details: {}
    type: object
  http.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the backend API for E-Commerce platform
  title: E-Commerce API
  version: "1.0"
paths:
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Request a password reset token to be sent to email
      parameters:
      - description: Forgot Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
      summary: Forgot password
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
      summary: Login user
      tags:
      - auth
  /auth/logout:
    post:
      description: Invalidate the current JWT token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with name, email, and password
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
      summary: Register a new user
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password using token
      parameters:
      - description: Reset Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
      summary: Reset password
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"